Windows CI integration for custom ops
