+++ test/custom_operator/CMakeLists.txt
@@ -5,7 +5,9 @@ project(custom_ops)
 find_package(Torch REQUIRED)
 
 add_library(custom_ops SHARED op.cpp)
-target_link_libraries(custom_ops ${TORCH_LIBRARIES})
+target_compile_features(custom_ops PUBLIC cxx_range_for)
+target_link_libraries(custom_ops "${TORCH_LIBRARIES}")
+target_compile_definitions(custom_ops PRIVATE custom_ops_EXPORTS)
 
 add_executable(test_custom_ops test_custom_ops.cpp)
 target_link_libraries(test_custom_ops custom_ops)
