+++ test/custom_operator/test_custom_ops.cpp
@@ -2,7 +2,6 @@
 
 #include "op.h"
 
-#include <cassert>
 #include <memory>
 #include <string>
 #include <vector>
@@ -26,10 +25,10 @@ void check_all_parameters(
 void get_operator_from_registry_and_execute() {
   auto& ops = torch::jit::getAllOperatorsFor(
       torch::jit::Symbol::fromQualString("custom::op"));
-  assert(ops.size() == 1);
+  AT_ASSERT(ops.size() == 1);
 
   auto& op = ops.front();
-  assert(op->schema().name == "custom::op");
+  AT_ASSERT(op->schema().name == "custom::op");
 
   torch::jit::Stack stack;
   torch::jit::push(stack, torch::ones(5), 2.0, 3);
@@ -39,10 +38,10 @@ void get_operator_from_registry_and_execute() {
 
   const auto manual = custom_op(torch::ones(5), 2.0, 3);
 
-  assert(output.size() == 3);
+  AT_ASSERT(output.size() == 3);
   for (size_t i = 0; i < output.size(); ++i) {
-    assert(output[i].allclose(torch::ones(5) * 2));
-    assert(output[i].allclose(manual[i]));
+    AT_ASSERT(output[i].allclose(torch::ones(5) * 2));
+    AT_ASSERT(output[i].allclose(manual[i]));
   }
 }
 
@@ -50,26 +49,26 @@ void load_serialized_module_with_custom_op_and_execute(
     const std::string& path_to_exported_script_module) {
   std::shared_ptr<torch::jit::script::Module> module =
       torch::jit::load(path_to_exported_script_module);
-  assert(module != nullptr);
+  AT_ASSERT(module != nullptr);
 
   std::vector<torch::jit::IValue> inputs;
   inputs.push_back(torch::ones(5));
   auto output = module->forward(inputs).toTensor();
 
-  assert(output.allclose(torch::ones(5) + 1));
+  AT_ASSERT(output.allclose(torch::ones(5) + 1));
 }
 
 void test_argument_checking_for_serialized_modules(
     const std::string& path_to_exported_script_module) {
   std::shared_ptr<torch::jit::script::Module> module =
       torch::jit::load(path_to_exported_script_module);
-  assert(module != nullptr);
+  AT_ASSERT(module != nullptr);
 
   try {
     module->forward({torch::jit::IValue(1), torch::jit::IValue(2)});
-    assert(false);
+    AT_ASSERT(false);
   } catch (const c10::Error& error) {
-    assert(
+    AT_ASSERT(
         std::string(error.what_without_backtrace())
             .find("Expected at most 1 argument(s) for operator 'forward', "
                   "but received 2 argument(s)") == 0);
@@ -77,9 +76,9 @@ void test_argument_checking_for_serialized_modules(
 
   try {
     module->forward({torch::jit::IValue(5)});
-    assert(false);
+    AT_ASSERT(false);
   } catch (const c10::Error& error) {
-    assert(
+    AT_ASSERT(
         std::string(error.what_without_backtrace())
             .find("Expected value of type Dynamic for argument 'input' in "
                   "position 0, but instead got value of type int") == 0);
@@ -87,9 +86,9 @@ void test_argument_checking_for_serialized_modules(
 
   try {
     module->forward({});
-    assert(false);
+    AT_ASSERT(false);
   } catch (const c10::Error& error) {
-    assert(
+    AT_ASSERT(
         std::string(error.what_without_backtrace())
             .find("forward() is missing value for argument 'input'") == 0);
   }
