+++ torch/csrc/jit/script/module.cpp
@@ -13,30 +13,6 @@ void placeholderCreator(Method&) {
   throw RecursiveMethodCallError();
 }
 
-static FunctionSchema defaultSchemaFor(const Method& method) {
-  std::vector<Argument> args;
-  std::vector<Argument> returns;
-  Graph& g = *method.graph();
-  size_t num_inputs = method.num_inputs();
-  for(size_t i = 0; i < num_inputs; ++i) {
-    const Value* v = g.inputs().at(i);
-    std::string name = v->hasUniqueName() ? v->uniqueName() : ("argument_"  + std::to_string(i));
-    args.push_back({std::move(name), unshapedType(g.inputs()[i]->type())});
-  }
-  for(size_t i = 0; i < g.outputs().size(); ++i) {
-    returns.push_back({"", unshapedType(g.outputs()[i]->type())});
-  }
-  return { method.name(), std::move(args), std::move(returns) };
-}
-
-
-const FunctionSchema& Method::getSchema() const {
-  if(schema == nullptr) {
-    schema.reset(new FunctionSchema(defaultSchemaFor(*this)));
-  }
-  return *schema;
-}
-
 c10::optional<std::vector<Value*>> try_emit_call_to(
     Graph& graph,
     SourceRange loc,
